Fix React rendering
React did a full rerender on state updates. Turns out this is because we create the conditional component in the render method each time it's renderer, so React doesn't have anything to diff against. This is a dangerous thing with stateless components, as it makes it too easy to do too much in the render function.

Closes #84