Fix stopping marketmaker on logout (#306)
When fixing an issue a long time ago (https://github.com/lukechilds/hyperdex/commit/dfda2d944380691b287ab82f0348eb385d556df8), I introduced a bug. The fix then was to pause the queue so we wouldn't get any zombie calls after logging out. The problem is that the API `stop` command also uses the queue, so it was never run. We now only pause the queue after the `stop` command is done.