Add `alwaysEnabledCurrencies` constant (#329)
This is required for ETH/ERC20, that's not ready yet but this is still an improvement over the current code so I've cherry picked it out from my local ETH branch.

It's nice to have this defined somewhere rather than hardcode it in a few places.

This also ensures these currencies are always activated (even if they're removed from the config). That's good because it'll break things if KMD (and soon ETOMIC) aren't enabled.

This also means we can easily enforce new currencies in updates (like ETOMIC). Updating the default coins in the config won't work after the initial install.