Fix Windows login issue
Windows seems to have a very low limit for the length of command arguments when spawning, so our recent currency additions made it go over that limit. A quick fix for now is to exclude the `electrumServers` property as it's not used by marketmaker. A proper longer-term fix would be for marketmaker to let us dynamically toggle coins with all this info instead of having to supply it on startup.

Fixes #519
Fixes #525